move Core::Render as inner class of Tilt
