+++ lib/munge.rb
@@ -16,7 +16,6 @@ require "munge/helper/rendering"
 require "munge/transformer/tilt"
 require "munge/readers/filesystem"
 require "munge/core/config"
-require "munge/core/renderer"
 require "munge/core/router"
 require "munge/core/item_factory"
 require "munge/core/item_factory/content_parser"
